requirements.txt
# Core Django stack
Django
djangorestframework
drf-yasg
requests

celery
redis

# Machine learning & data processing
scikit-learn
pandas
numpy
joblib

# Telegram bot integration
python-telegram-bot

# Database
psycopg2-binary  # Use only in development; use `psycopg2` in production

# Deployment
gunicorn

# Optional: Use whitenoise to serve static files in production
whitenoise


# Google backup
gcs-oauth2-boto-plugin
gdown
google-api-core>=2.11.0
google-api-python-client>=2.100.0
google-apitools
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-cloud-core>=2.3.2
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
google-crc32c>=1.5.0
google-pasta>=0.2.0
google-reauth>=0.1.1
google-resumable-media>=2.6.0
googleapis-common-protos>=1.59.1
google-auth-oauthlib>=1.2.0   # 🔑 ensures run_console() exists

aiohttp>=3.9.5